+++ Studies/include/MvaVariablesStudy.h
@@ -12,7 +12,7 @@ namespace mva_study{
 using DataVector = std::vector<double>;
 using VarData = std::map<std::string, DataVector>;
 using SampleIdVarData = std::map<SampleId, VarData>;
-using ChannelSpin = std::pair<std::string,double>;
+using ChannelSpin = std::pair<std::string,int>;
 
 
 class MvaVariablesStudy : public MvaVariables {
@@ -28,7 +28,7 @@ public:
         variables[name] = value;
     }
 
-    virtual void AddEventVariables(size_t set, const SampleId& mass,  double /*weight*/, double /*sampleweight*/, double spin, std::string channel) override
+    virtual void AddEventVariables(size_t set, const SampleId& mass,  double /*weight*/, double /*sampleweight*/, int spin, std::string channel) override
     {
         ChannelSpin chsp(channel,spin);
         VarData& sample_vars = all_variables[chsp][set][mass];
@@ -39,7 +39,7 @@ public:
         }
     }
 
-    const SampleIdVarData& GetSampleVariables(std::string channel, double spin, size_t set = 0) const
+    const SampleIdVarData& GetSampleVariables(std::string channel, int spin, size_t set = 0) const
     {
         if(set >= all_variables.size())
             throw exception("Sample part is out of range.");
